requirements.txt
cycler==0.10.0
Django==2.0.1
django-cors-headers==2.1.0
django-extensions==1.9.9
imutils==0.4.5
matplotlib==2.1.2
numpy==1.13.3
opencv-python==3.3.1.11
pandas==0.22.0
Pillow==5.0.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
typing==3.6.4
